% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transph.R
\name{confint.transph}
\alias{confint.transph}
\title{Confidence intervals for estimated coefficients}
\usage{
\method{confint}{transph}(creg, parm, level = 0.95, type = "wald")
}
\arguments{
\item{parm}{A coefficient name or vector of coefficient names. If missing, #'  confidence intervals are calculated for all estimated coefficients.}

\item{level}{The confidence level (1 - \eqn{alpha}).}

\item{type}{The type of confidence interval. Current options are \code{wald} 
for Wald confidence limits and \code{lr} for likelihood ratio confidence 
limits. When who-infected-whom is not completely observed, the latter are 
under development and should not be used in practice.}

\item{treg}{An object of class \code{transph}.}
}
\value{
A data frame with one row for each coefficient in \code{parm} that 
 contains the lower and upper confidence limits in columns labeled 
 \eqn{\frac{\alpha}{2}} and \eqn{1 - \frac{\alpha}{2}} expressed as 
 percentiles.
}
\description{
Calculates confidence intervals for coefficient estimates from a 
\code{transph} model by inverting the Wald or likelihood ratio tests.
}
\author{
Eben Kenah \email{kenah.1@osu.edu}
}
