% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transreg.R
\name{confint.transreg}
\alias{confint.transreg}
\title{Confidence intervals for estimated parameters}
\usage{
\method{confint}{transreg}(treg, parm, level = 0.95, method = "wald")
}
\arguments{
\item{treg}{An object of class \code{treg}.}

\item{parm}{A parameter or vector of parameters. If missing, confidence 
intervals are calculated for all estimated parameters.}

\item{level}{The confidence level (1 - \eqn{alpha}).}

\item{method}{The type of confidence interval. Current options are 
\code{wald} for Wald confidence limits and \code{lr} for likelihood ratio
confidence limits.}
}
\value{
A data frame containing the lower and upper confidence limits with 
 column labeled with \eqn{\frac{\alpha}{2}} and 
 \eqn{1 - \frac{\alpha}{2}} expressed as percentages.
}
\description{
Calculates confidence intervals for estimated parameters from a 
\code{transreg} model by inverting the Wald or likelihood ratio tests.
}
\author{
Eben Kenah \email{ekenah@ufl.edu}
}
