% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ext}
\alias{ext}
\title{Identify rows corresponding to external contact intervals.}
\usage{
ext(x)
}
\arguments{
\item{ex}{A binary variable such that \code{as.numeric(ex)} will produce
a one for each external row and a zero for each internal row}
}
\value{
A factor with two levels
}
\description{
Specifies indicator variable for rows containing an external contact
interval (from the community or environment to a susceptible individual)
in model formula.
}
\details{
If \code{ex} is an indicator variable for external infection, then
 \code{exog(ex)} can be added to the model formula to allow separate
 handling of internal and external contact intervals.
}
