% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transreg.R
\name{print.transreg_summary}
\alias{print.transreg_summary}
\title{Print summary of fitted parametric regression model for ID transmission}
\usage{
\method{print}{transreg_summary}(treg_sum, cdigits = 4, pdigits = 3)
}
\arguments{
\item{treg_sum}{An object of class \code{transreg_summary}.}

\item{cdigits}{The minimum number of significant digits to print for 
parameter point and interval estimates.}

\item{pdigits}{The minimum number of significant digits to print for p-
values. This is passed to \code{\link[base]{format.pval}}.}
}
\description{
Prints a summary of a fitted \code{transreg} model. The number of digits
to print for parameter estimates and p-values can be specified. The 
p-values are formatted using \code{\link[base]{format.pval}}.
}
\author{
Eben Kenah \email{ekenah@ufl.edu}
}
