% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pval}
\alias{pval}
\title{p-values for model parameters.}
\usage{
pval(object, parm, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{parm}{A parameter or vector of parameters. If missing, p-values are
calculated for all estimated parameters.}

\item{...}{Additional arguments for methods.}
}
\value{
A named vector of p-values.
}
\description{
Calculates p-values for one or more estimated parameters.
}
\details{
\code{pval} is a generic function. The default method calculates Wald
 confidence intervals and assumes that \code{object} has \code{coef} and
 \code{vcov} methods available.
}
