% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transph.R
\name{pval.transph}
\alias{pval.transph}
\title{p-values for estimated coefficients}
\usage{
\method{pval}{transph}(creg, parm, type = "wald")
}
\arguments{
\item{parm}{A coefficient name or vector of coefficient names. If missing, 
p-values are calculated for all estimated parameters.}

\item{type}{The type of p-value. Options are \code{wald} for Wald p-values 
and \code{lr} for likelihood ratio p-values. Likelihood ratio p-values are 
available only when who-infected-whom is completely observed.}

\item{treg}{An object of class \code{transph}.}
}
\value{
A named vector of p-values.
}
\description{
Calculates p-values for coefficient estimates from a \code{transph} model 
using a normal approximation or a likelihood ratio chi-squared statistic.
}
\author{
Eben Kenah \email{kenah.1@osu.edu}
}
