% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transph.R
\name{survfit.transph}
\alias{survfit.transph}
\title{Compute survival and cumulative hazard curves from a transph model}
\usage{
survfit.transph(creg, newdata, level = 0.95, transform = "log-log", ...)
}
\arguments{
\item{newdata}{A data frame or named vector with the same variable names 
as the \code{transph} formula. It will be passed to 
\code{\link[survival]{survfit.coxph}}.}

\item{level}{The level for pointwise two-sided confidence limits.}

\item{transform}{The transformation of the survival function used to 
calculate confidence limits. The "log-log" option uses the log cumulative 
hazard, and the "log" option uses the cumulative hazard directly.}

\item{...}{Further arguments to be passed to 
\code{\link[survival]{survfit.coxph}}, such as \code{type}, \code{vartype}, 
and \code{start.time}. Do not include \code{se.fit}, which is \code{TRUE}
(the default), or \code{conf.type}, which is set to \code{none}.}

\item{treg}{A \code{transph} object.}
}
\value{
A \code{\link[survival]{survfit.object}} with its \code{std.err}, 
 \code{upper}, and \code{lower} components altered as needed to account for 
 uncertainty in who-infected-whom. The \code{conf.int} and \code{conf.type} 
 components are altered to match the \code{level} and \code{transform} 
 arguments, respectively. This object can be passed to the 
 following methods from the \code{survival} package:
 \code{\link[survival]{plot.survfit}}, 
 \code{\link[survival]{summary.survfit}}, and 
 \code{\link[survival]{print.survfit}}.
}
\description{
Compute survival and cumulative hazard curves from a transph model
}
\author{
Eben Kenah \email{kenah.1@osu.edu}
}
