% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{transfit}
\alias{transfit}
\title{Nonparametric estimators for infectious disease transmission}
\usage{
transfit(
  formula,
  sus,
  data,
  subset = NULL,
  weights,
  na.action,
  inithaz = NULL,
  maxit = 50
)
}
\arguments{
\item{formula}{A formula of the form "response ~ terms". The response must 
be an object returned by \code{\link[survival]{Surv}}. The terms are 
converted into a factor with a level corresponding to each unique 
combination of covariate values.}

\item{sus}{A string giving the name of the variable indicating the 
susceptible member of each pair.}

\item{data}{A data frame containing the variables named in \code{formula}.}

\item{subset}{An expression indicating which rows of \code{data} should be
included in the model fit. Default is to keep all rows.}

\item{na.action}{A missing-data filter applied to \code{data} after
\code{subset}. Defaults to \code{options()$na.action}.}

\item{inithaz}{A named hazard function (up to a constant of proportionality) #'  to be used in the first iteration. The formal arguments in the function 
need to match variable names \code{data} that appear in \code{formula}.}

\item{maxit}{The maximum number of EM algorithm iterations.}
}
\description{
Calculates Nelson-Aalen or Kaplan-Meier curves for infectious disease 
transmission using right-censored and/or left truncated data on contact 
intervals in ordered pairs of individuals and infectious contacts from 
external sources with individuals.
}
\references{
E Kenah (2013). Nonparametric survival analysis of infectious 
 disease data. \emph{Journal of the Royal Statistical Society, Series B} 
 75(2): 277-303.
}
\author{
Eben Kenah \email{kenah.1@osu.edu}
}
