% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transph.R
\name{transph}
\alias{transph}
\title{Semiparametric regression models for infectious disease transmission}
\usage{
transph(
  formula,
  sus,
  data,
  weights,
  subset,
  na.action,
  spline_df,
  itermax = 25,
  L1tol = 1e-04,
  ...
)
}
\arguments{
\item{formula}{A formula of the form "response ~ terms". The response 
must be an object returned by \code{\link[survival]{Surv}} of type 
"right" or "counting". The formula can include \code{offset} terms for 
fixed coefficients, and an external row indicator can be included in a 
\code{strata} term.}

\item{sus}{A string giving the name of the variable in \code{data} that
contains the susceptible member of each pair.}

\item{data}{A data frame containing the variables named in \code{formula}.}

\item{weights}{A vector of infector probabilities for the data rows. If 
who-infected-whom is not completely observed, these will be updated in an
EM algorithm. If missing, all possible infectors of each susceptible are
given equal initial weights. If who-infected-whom is observed, the weights 
are passed to \code{\link[survival]{coxph}}, where they will be used as 
case weights. Normally, there would be no weights in this case.}

\item{subset}{An expression indicating which rows of \code{data} should be
included in the model fit.}

\item{na.action}{A missing-data filter applied to \code{data} after
\code{subset}. Defaults to \code{options()$na.action}.}

\item{itermax}{The maximum number of EM algorithm iterations if who-
infected-whom is not completely observed. Setting \code{itermax = 1} will 
cause the model to to use only the initial weights (default or provided).}

\item{L1tol}{The EM algorithm stops when the L1 distance between the old and 
new weights is below \code{L1tol} per transmission event.}

\item{...}{Further arguments to be passed to \code{\link[survival]{coxph}},
such as \code{init}, \code{ties}, or \code{control}.}

\item{method}{The method used to generate a monotonic cubic interpolating 
spline for the cumulative hazard. Options are \code{monoH.FC} and 
\code{hyman}. See documentation for \code{\link[splines]{splinefun}}.}
}
\value{
A list of class \code{transph} with the following components:
 \describe{
   \item{\code{call}}{The call to \code{transph} with complete formal 
     arguments.}
   \item{\code{coefficients}}{The estimated coefficients.}
   \item{\code{coxph}}{The \code{\link[survival]{coxph}} object returned by 
     the final Cox regression.}
   \item{\code{data}}{The data used to fit the model.}
   \item{\code{df}}{The number of estimated coefficients.}
   \item{\code{formula}}{The model formula.}
   \item{\code{iter}}{The number of iterations. It is one for a model in 
     which who-infects-whom is completely observed. Otherwise, it is the 
     number of EM algorithm iterations (including the initial model).}
   \item{\code{L1tol}}{The L1 tolerance per transmission event used to halt 
     the EM algorithm.}
   \item{\code{loglik}}{The \code{loglik} element in the 
     \code{\link[survival]{coxph.object}} returned by the final Cox 
     regression, which contains the log partial likelihood at the 
     initial coefficient values and the fitted coefficient values. For a 
     null model, only the first element is present. This is \code{NA} when 
     who-infected-whom is not completely observed.}
   \item{\code{smooth_method}}{The \code{method} element in the 
     \code{\link[survival]{coxph.object}} returned by the final Cox 
     regression, which is the name of approximation used to handle ties.}
   \item{\code{spline_df}}{The degrees of freedom in the smoothing spline 
     used to calculate the contact interval hazard function(s).}
   \item{\code{sus}}{The vector identifying the susceptible member of each #'      pair in the data used to fit the model.}
   \item{\code{var}}{The covariance matrix for the coefficient estimates. #'      The rows and columns are named for the corresponding covariates.}
   \item{\code{weights}}{The final vector of estimated infector 
     probabilities.}
   \item{\code{ymat}}{The outcome matrix returned by 
     \code{\link[survival]{Surv}}}.
 }
}
\description{
Fits semiparametric pairwise regression models for infectious disease   
transmission using right-censored and/or left-truncated data on contact 
intervals in ordered pairs of individuals and infectious contact from 
external sources with individuals. Uses the Cox relative risk function. 
External sources of infection are handled by stratifying on an external 
row indicator.
}
\author{
Eben Kenah \email{kenah.1@osu.edu}
}
