% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transreg.R
\name{transreg}
\alias{transreg}
\title{Parametric regression models for infectious disease transmission}
\usage{
transreg(formula, data, subset = NULL, na.action, sus, dist = "weibull",
  init = NULL, fixed = NULL, optim_method = "BFGS", ...)
}
\arguments{
\item{formula}{A formula of the form "response ~ terms". The response 
must be an object returned by \code{\link[survival]{Surv}}. Only
right-censored data is supported.}

\item{data}{A data frame containing the variables named in \code{formula}.}

\item{subset}{An expression indicating which rows of \code{data} should be
included in the model fit.}

\item{na.action}{A missing-data filter applied to \code{data} after
\code{subset}. Defaults to \code{options()$na.action}.}

\item{sus}{A string giving the name of the variable in \code{data} that
contains the susceptible member of each pair.}

\item{dist}{A string partially matching a survival time distribution from 
\code{\link{transreg.distributions}}. Current options are 
\code{exponential}, \code{weibull} (the default), and \code{loglogistic}.}

\item{init}{A named vector of initial values for estimated coefficients.}

\item{fixed}{A named vector of fixed parameter values. These can include 
terms in \code{formula}, \code{(Intercept)}, and \code{log(shape)}.}

\item{optim_method}{The method to be used by \code{\link[stats]{optim}}.}

\item{...}{Further arguments to be passed to \code{\link[stats]{optim}}.}
}
\value{
A list with class \code{transreg} that contains the following 
 objects:
 \describe{
   \item{\code{call}}{The call to \code{transreg} with complete formal 
     arguments.}
   \item{\code{coef}}{Named vector of estimated parameters.}
   \item{\code{df}}{The number of estimated coefficients.}
   \item{\code{dist}}{String naming the survival time distribution.}
   \item{\code{fixed}}{Named vector of fixed parameters.}
   \item{\code{loglik}}{The maximum log likelihood.}
   \item{\code{model_matrix}}{The data frame used to fit the model.}
   \item{\code{nlnL}}{Function for calculating the negative log 
     likelihood. See details below.}
   \item{\code{optim_method}}{The method used in 
     \code{\link[stats]{optim}}.}
   \item{\code{response}}{The response from \code{\link[survival]{Surv}}.}
   \item{\code{sus}}{Factor giving the susceptible member of each 
     ordered pair.}
   \item{\code{var}}{The estimated variance matrix.}
 }
}
\description{
Fits accelerated failure time models for infectious disease transmission 
data using right-censored data on contact intervals in ordered pairs of 
individuals. Who-infected-whom can be observed, unobserved, or partially 
observed.
}
\details{
\strong{Survival time distributions}
 \itemize{
   \item {The exponential distribution (\code{exponential}) has a rate 
     parameter \eqn{\lambda}. Its hazard function is 
     \deqn{h(t, \lambda) = \lambda.}
   }
   \item{The Weibull distribution (\code{weibull}) has a rate parameter 
     \eqn{\lambda} and a shape parameter \eqn{\gamma}. Its cumulative 
     hazard function is 
     \deqn{H(t, \lambda, \gamma) = (\lambda t)^\gamma.}
     The exponential distribution is a Weibull distribution with 
     shape \eqn{\gamma = 1}.
   }
   \item{The log-logistic distribution (\code{loglogistic}) has a rate 
     parameter \eqn{lambda} and a shape parameter \eqn{\gamma}. Its 
     survival function is
     \deqn{S(t, \lambda, \gamma) = \frac{1}{1 + (\lambda t)^\gamma}.}
   }
 }
 \strong{Accelerated failure time models} are log-linear models for the 
 rate parameter \eqn{\lambda} in the specified survival time distribution. 
 Each coefficient can be interpreted as the log rate ratio for a one-unit 
 increase in the corresponding covariate, so each covariate has a 
 multiplicative effect on the rate. A single shape parameter \eqn{\gamma}
 is estimated for each stratum.
}
\references{
E Kenah (2011). Contact intervals, survival analysis of 
 epidemic data, and estimation of R_0. \emph{Biostatistics} 12(3): 
 548-566.

E Kenah and Y Sharker (2016)
}
\author{
Eben Kenah \email{ekenah@ufl.edu}
}
