% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transreg.R
\name{confint.transreg}
\alias{confint.transreg}
\title{Confidence intervals for estimated parameters}
\usage{
\method{confint}{transreg}(
  treg,
  parm,
  level = 0.95,
  type = "wald",
  lrint = c(0.2, 5),
  extendInt_lower = "downX",
  extendInt_upper = "upX",
  ...
)
}
\arguments{
\item{treg}{An object of class \code{transeg}.}

\item{parm}{A parameter name or vector of parameter names. If missing,
confidence intervals are calculated for all estimated parameters.}

\item{level}{The confidence level (1 - \eqn{alpha}).}

\item{type}{The type of confidence interval. Current options are \code{wald}
for Wald confidence limits and \code{lr} for likelihood ratio confidence
limits. The latter are more accurate but more computationally intensive.}

\item{lrint}{The highest and lowest multiples of the standard error used by
\code{\link[stats]{uniroot}} to find likelihood ratio confidence limits.}

\item{extendInt_lower}{The extendInt argument in \code{\link[stats]{uniroot}}
when calculating the lower likelihood ratio confidence limit.}

\item{extendInt_upper}{The extendInt argument in \code{\link[stats]{uniroot}}
when calculating the upper likelihood ratio confidence limit.}

\item{...}{Additional arguments sent to \code{\link[stats]{uniroot}}.}
}
\value{
A data frame with one row for each parameter in \code{parm} that
 contains the lower and upper confidence limits in columns labeled
 \eqn{\frac{\alpha}{2}} and \eqn{1 - \frac{\alpha}{2}} expressed as
 percentiles.
}
\description{
Calculates confidence intervals for estimated parameters from a
\code{transreg} model by inverting the Wald or likelihood ratio tests.
}
\author{
Eben Kenah \email{kenah.1@osu.edu}
}
