% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transph.R
\name{transph}
\alias{transph}
\title{Semiparametric regression models for infectious disease transmission}
\usage{
transph(formula, sus, data, weights = NULL, subset = NULL, na.action,
  itermax = 25, degf = NULL, L1tol = 1e-04, ...)
}
\arguments{
\item{formula}{A formula of the form "response ~ terms". The response 
must be an object returned by \code{\link[survival]{Surv}} of type "right" 
or "counting". The formula can include \code{offset} terms for fixed 
coefficients, and the external row indicator must be included in a 
\code{strata} term.}

\item{sus}{A string giving the name of the variable in \code{data} that
contains the susceptible member of each pair.}

\item{data}{A data frame containing the variables named in \code{formula}.}

\item{weights}{A vector of infector probabilities for the data rows. If 
who-infected-whom is not completely observed, these will be updated in an
EM algorithm. If missing, all possible infectors of each susceptible are
given equal initial weights.}

\item{subset}{An expression indicating which rows of \code{data} should be
included in the model fit.}

\item{na.action}{A missing-data filter applied to \code{data} after
\code{subset}. Defaults to \code{options()$na.action}.}

\item{itermax}{The maximum number of EM algorithm iterations if who-
infected-whom is not observed.}

\item{degf}{The degrees of freedom for the smoothing spline used to obtain 
the hazard function(s) from the estimated cumulative hazard(s). The default 
is to use twice the natural logarithm of the total number of transmissions.}

\item{L1tol}{The EM algorithm stops when the L1 distance between the old and 
new weights is below \code{L1tol} per transmission event.}

\item{...}{Further arguments to be passed to \code{\link[survival]{coxph}},
such as \code{init}, \code{ties}, and \code{control}.}
}
\description{
Fits semiparametric pairwise regression models for infectious disease   
transmission using right-censored and/or left-truncated data on contact 
intervals in ordered pairs of individuals and infectious contact from 
external sources with individuals. Uses the Cox relative risk function.
}
